data/retrieval/RAG/assets/rag-accelerator/requirements.txt
# Core Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
starlette==0.50.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-multipart==0.0.20
ipywidgets==8.1.8

# IBM Cloud Services
ibm-cloud-sdk-core==3.24.2
ibm-cos-sdk==2.14.3
ibm-cos-sdk-core==2.14.3
ibm-cos-sdk-s3transfer==2.14.3
ibm-watson==11.1.0
ibm_watsonx_ai==1.3.38

# Vector Databases
pymilvus==2.5.11
opensearch-py>=2.7.0
milvus-lite==2.5.1

# LangChain Ecosystem
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.80
langchain-elasticsearch==0.3.2
langchain-ibm==0.3.20
langchain-milvus==0.2.0
langchain-text-splitters==0.3.11
langsmith==0.4.56

# Document Processing
unstructured[all-docs]==0.16.0
pypdf==6.4.1
python-docx==1.1.2
python-pptx==1.0.2
beautifulsoup4==4.14.3
lxml==6.0.2

# HTTP & Async
aiohttp==3.13.2
httpx==0.28.1
httpx-sse==0.4.3
requests==2.32.5
requests-toolbelt==1.0.0

# Data Processing
numpy==1.26.4
pandas==2.2.3
nltk==3.9.2

# Utilities
python-dotenv==1.2.1
tqdm==4.67.1
PyYAML==6.0.3
click==8.3.1
tenacity==9.1.2

# Security & Authentication
PyJWT==2.10.1
certifi==2025.11.12

# Database & Storage
SQLAlchemy==2.0.44
elasticsearch==8.18.1

# Optional: Development & Testing (comment out for production)
# pytest==8.3.0
# pytest-asyncio==0.24.0
# pytest-cov==6.0.0
# black==24.10.0
# flake8==7.1.0
# mypy==1.13.0
